# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/families/complete-registration": {
      "post": {
        "operationId": "FamiliesController_completeRegistration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteFamilyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
    "/api/v1/families/me": {
      "get": {
        "operationId": "FamiliesController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Families"]
      },
      "patch": {
        "operationId": "FamiliesController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
    "/api/v1/families/addresses": {
      "post": {
        "operationId": "FamiliesController_addAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
    "/api/v1/families/addresses/{addressId}": {
      "patch": {
        "operationId": "FamiliesController_updateAddress",
        "parameters": [
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressMutationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
    "/api/v1/families/emergency-contacts/{contactId}": {
      "patch": {
        "operationId": "FamiliesController_updateEmergencyContact",
        "parameters": [
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyMutationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
    "/api/v1/families/set-primary-phone": {
      "post": {
        "operationId": "FamiliesController_setPrimaryPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParentTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
    "/api/v1/families/change-primary-phone": {
      "post": {
        "operationId": "FamiliesController_changePrimaryPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParentTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Families"]
      }
    },
...
    "/api/v1/admin/families/{id}": {
      "get": {
        "operationId": "AdminFamiliesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      }
    },
    "/api/v1/admin/families/{id}/parents": {
      "post": {
        "operationId": "AdminFamiliesController_createParent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateParentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      }
    },
    "/api/v1/admin/families/{id}/parents/{parentId}": {
      "patch": {
        "operationId": "AdminFamiliesController_updateParent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateParentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      },
      "delete": {
        "operationId": "AdminFamiliesController_deleteParent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      }
    },
    "/api/v1/admin/families/{id}/addresses": {
      "post": {
        "operationId": "AdminFamiliesController_addAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      }
    },
    "/api/v1/admin/families/{id}/addresses/{addressId}": {
      "patch": {
        "operationId": "AdminFamiliesController_updateAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressMutationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      }
    },
    "/api/v1/admin/families/{id}/emergency-contacts/{contactId}": {
      "patch": {
        "operationId": "AdminFamiliesController_updateEmergencyContact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyMutationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["AdminFamilies"]
      }
    },
...
  },
  "info": {
    "title": "School Transport API",
    "description": "Canonical REST contract for the school transport MVP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CompleteFamilyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
      "AddAddressDto": {
        "type": "object",
        "properties": {}
      },
      "AddressMutationDto": {
        "type": "object",
        "properties": {}
      },
      "EmergencyMutationDto": {
        "type": "object",
        "properties": {}
      },
      "ParentTypeDto": {
        "type": "object",
        "properties": {}
      },
      "AdminCreateParentDto": {
        "type": "object",
        "properties": {}
      },
      "AdminUpdateParentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}